alx_travel_app/requirement.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-yasg==1.21.8
python-decouple==3.8

# Celery and message broker dependencies
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Optional: For monitoring Celery tasks
flower==2.0.1

# Additional email dependencies
django-extensions==3.2.3